+++ packages/tsurlfilter/CHANGELOG.md
@@ -10,12 +10,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support of `$header` modifier in declarative converter.
+- Support of `:has()` pseudo-class as standard CSS conditionally:
+    - if there are no other extended pseudo-classes in the selector, or
+    - if the `#?#` separator is explicitly used in the rule [AdguardBrowserExtension#2587].
+
+### Changed
+
+<!-- TODO: update the version later -->
+- Updated [@adguard/agtree] to `vX.X.X`.
 
 ### Fixed
 
 - Apply $document blocking rule even if specific exception rule is present [AdguardBrowserExtension#3262].
 - The extension got an unexpected error on `[::]:8000` [AdguardBrowserExtension#3360].
 
+[AdguardBrowserExtension#2587]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2587
 [AdguardBrowserExtension#3262]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3262
 [AdguardBrowserExtension#3360]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3360